Organizing subsystems into separate sub-packages
