+++ src/main/java/frc/robot/commands/ElevatorTestCommand.java
@@ -5,7 +5,7 @@ import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.kinematics.SwerveModuleState;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.Subsystem;
-import frc.robot.subsystems.ElevatorSubsystem;
+import frc.robot.subsystems.score.ElevatorSubsystem;
 
 
 public class ElevatorTestCommand extends Command {
